BackendApplication/requirements.txt
flask==3.0.3
pymongo==4.7.2
pydantic==2.8.2
pyyaml==6.0.2
